+++ database/migrations/2023_02_23_085444_create_sallers_table.php
@@ -21,7 +21,7 @@ class CreateSallersTable extends Migration
         $table->string('type');
         $table->String('address');
         $table->integer('mobile')->nullable();
-        $table->integer('status')->default(0);
+        $table->tinyInteger('status')->default(0);
         $table->timestamps();
         $table->softDeletes();
       });
